requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
google-genai>=0.8.0
python-dotenv>=1.0.0
playwright>=1.44.0
Pillow>=10.0.0
flask>=3.0.0
